+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Deprecated
 ### Removed
 ### Fixed
+- Fixed double encoding of index ([#348](https://github.com/opensearch-project/opensearch-php/pull/348))
 ### Security
 ### Updated APIs